[4, 5단계 - 체스] 알리(허진호) 미션 제출합니다. (#231)
* feat : 체스 말 이미지 리소스 추가

* feat : database connection 추가

* refactor : 패키지 변경

* refactor : 패키지 변경에 따른 수정 사항

* feat : User 객체와 DB에 저장하기 위한 UserDao 생성

* feat : Game 객체와 DB에 저장하기 위한 GameDao 생성

* feat : GameHistory 객체와 DB에 저장하기 위한 GameHistoryDao 생성

* feat : Service 객체 추가

* feat : web을 위한 controller 추가

* feat : 로그인/게임방/체스게임 화면 구현

* fix : 대각선 이동 시 상대 말을 먹지 못하는 버그 수정

* polish : 코드정리

* refactor : db connection 오류 발생 시 프로그램 종료

* refactor : db 조회 후 조회 결과가 없을 경우 null 대신 Optional 반환

* refactor : 잘못된 Null 체크 삭제

* refactor : 로그인 확인이 필요한 함수들을 optional을 사용해서 사용자 쿠키를 확인하고 사용자 없는 경우 redirect 처리

+++ src/main/java/chess/domain/position/Position.java
@@ -1,7 +1,7 @@
 package chess.domain.position;
 
 import chess.domain.piece.Direction;
-import chess.domain.util.BoardInitializer;
+import chess.util.BoardInitializer;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -79,8 +79,8 @@ public class Position {
             Rank.of(this.rank.getRank() + dir.getYDegree() * step));
     }
 
-    public int calculateDistance(Position source) {
-        return (int) Math.sqrt(
+    public double calculateDistance(Position source) {
+        return Math.sqrt(
             (Math.pow(this.getX() - source.getX(), 2) + Math.pow(this.getY() - source.getY(), 2)));
     }
